requirements.txt
apiclient==1.0.4
backports.csv==1.0.7
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==1.11.17
django-all-access==0.9.0
django-import-export==1.2.0
django-nested-admin==3.2.3
djangorestframework==3.10.2
et-xmlfile==1.0.1
google-api-python-client==1.7.10
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
httplib2==0.13.1
httplib2shim==0.0.3
idna==2.8
jdcal==1.4.1
jsonfield==2.0.2
oauth2client==4.1.3
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==2.6.2
psycopg2==2.8.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycrypto==2.6.1
python-dateutil==2.8.0
python-monkey-business==1.0.0
pytz==2019.2
PyYAML==5.1.1
rauth==0.7.3
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
tablib==0.13.0
uritemplate==3.0.0
urllib3==1.25.3
xlrd==1.2.0
xlwt==1.3.0
